requirements.txt
flask
groq
requests
python-dotenv
Pillow
gunicorn
fastapi
uvicorn
